:recycle: move tests in a dedicated file
