+++ marketplace-core/src/application/contribution/refresh/mod.rs
@@ -1,0 +1,6 @@
+use crate::application::refresh::Refresh;
+use marketplace_domain::*;
+
+pub type RefreshContributions = Refresh<ContributionProjection, Contribution>;
+pub type RefreshApplications = Refresh<ApplicationProjection, Contribution>;
+pub type RefreshProjects = Refresh<ProjectProjection, Contribution>;
