+++ marketplace-domain/src/contribution/projectors/project/mod.rs
@@ -1,0 +1,5 @@
+mod project;
+pub use project::*;
+
+#[cfg(test)]
+mod test;
